pyproject.toml
[tool.poetry]
name = "openrelik-worker-template"
version = "2024.11.27"
description = "Description of the worker"
authors = ["YOUR NAME <email@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
celery = {extras = ["redis"], version = "^5.4.0"}
openrelik-worker-common = "^2024.11.27"
google-cloud-bigquery = "^3.29.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
